YTD v1.16 (2012-09-23)
CHANGELOG
---------
v1.16:
	* New providers:
		- Azet.sk
		- EuroSeptik.cz
			- Subtitle download is also supported. By default the english subtitles
			  will be downloaded, but you can change it in the configuration by
			  using e.g. "Přepis, angličtina" or "Přepis, čeština" (the text can
			  be found on the video's page; you don't need to provide the whole
			  text, an identifying part such as "čeština" will be fine).
		- Matrix-2001.cz
		- MetroProjekt.cz
		- T-Music.cz (thanks to Balutbj)
		- Vikend.Nova.cz (and likely the other Nova.cz subdomains, too)
	* Updated providers:
		- FunnyOrDie.com
		- YouTube.com
	* Removed support for Huste.tv (and Huste.sk), because it requires a login
	  and I don't have an account there. If I had an account to play with, the
	  support would probably return pretty quickly.
	* Fixed the file deletion issue on invalid downloads - until now, empty
	  files only got deleted on "success", but not on actual errors such as
	  YouTube's 403 Forbidden.
	* Fixed the filenames generated for RTMP and MS protocols in the unicode
	  version. This fixes the random errors while downloading from ceskatelevize.cz,
	  which would work every time in the ANSI version.
	* RTMPDUMP updated to version 2012-09-22. The main feature of this version
	  is the realtime mode, which sacrifices some speed but greatly improves
	  the downloaded file's stability: In the normal mode, RTMPDUMP frequently
	  pauses and unpauses the playback, which forces the server to send a large
	  chunk of data without bandwidth limitation (which improves speed), but
	  unfortunately some servers skip back a few seconds on unpause, which
	  makes the downloaded movie virtually unusable. Realtime mode is turned on
	  by default for all providers, but you can turn it off (to get better
	  speed) in the provider's configuration.
	* Documentation converted to Unicode enconding.

+++ source/Base/uHttpDownloader.pas
@@ -148,6 +148,8 @@ begin
 end;
 
 function THttpDownloader.Download: boolean;
+const
+  MINIMUM_SIZE_TO_KEEP = 10240;
 var
   FN, FinalFN: string;
 begin
@@ -192,13 +194,14 @@ begin
             VideoDownloader.OutputStream.Free;
             VideoDownloader.OutputStream := nil;
             if Result then
-              if (not FileExists(FN)) or (GetFileSize(FN) <= 1024) then
+              if (not FileExists(FN)) or (GetFileSize(FN) < MINIMUM_SIZE_TO_KEEP) then
                 begin
-                if FileExists(FN) then
-                  DeleteFile(PChar(FN));
                 SetLastErrorMsg(ERR_HTTP_NO_DATA_READ);
                 Result := False;
                 end;
+            if not Result then
+              if FileExists(FN) and (GetFileSize(FN) < MINIMUM_SIZE_TO_KEEP) then
+                DeleteFile(PChar(FN));
             if Result then
               if FN <> FinalFN then
                 begin
